chore: Replace `Chat` by the `WKWebView` wrapper
